WIP - Main app flow improvements & logging, and missing schema descs
This is probably not a fully-working commit, particularly the EmptyDataModel class. But I want to try it on a new pipeline, and to do that I want to set up a new app repo, and to do that I want to write the script to automate setup.

this commit includes pushing full conf (confusingly) into conf.data init()