Fixed bug: stop getDataFromSrc removing first part of table name
An unexpected downstream impact of having two tables in two source systems with the same name, that need processing in the same dataflow. You have to specify a srcTableName and a tableName in the getDataFromSrc function, which means BETL thinks you're calling for a default extract, and so tries to remove the source system prefix.

Very hacky solution - no time to fix now.